Rework color generation to favor rgba values when `alpha < 1` (#406)
* Rework color gen to favor rgba when alpha < 1

* Return the hexadecimal color value

* Update help files with new version of roxygen2

* Rework color processing functions

This eliminates several of the color utility functions and adds `farver` as a dependency (which does a lot of the heavy lifting for color conversions). The `normalize_color()` function has been replaced with `process_color()`.

* Make fixes for passing R CMD check

Adds new items to the `globalVariables()` vector. Removes an unnecessary part of an example that isn't written to the .Rd file in the same way as before (with new roxygen2 version).

* Refactor color processing stmts in `data_color()`

* Add changes based on code review comments

* Add several tests for the `data_color()` function

* Modify docs for `data_color()`

These are corrections for the `alpha` argument

* Add several testthat tests

* Add several testthat tests for color handling

* Remove the dependency on farver; add several tests

* Make correction to example with invalid color name

* Modify the default for `alpha` in `cell_fill()`

This change allows for the use of #RRGGBBAA color values and `alpha` override in that case and add alpha information for all other color inputs. Documentation has been updated, and tests have been added.

* Refactor `data_color()`

* Fix bug where discrete data coloring uses interpolation unnecessarily