Fix unintended regression from dplyr group default
The change we made to groupname_col default caused
the groupnames to be set to "", not NA. This made
tables without groups render incorrectly. I've
fixed this issue and reverted the test changes that
were introduced yesterday to work around the change.

Also added some tests to test_fmt_number that ensure
group/rowname columns don't count as targetable cols
yet are visible in the data mask.