Squelch warnings from tibble 3.0.0/3.0.1 (#557)
* Avoid subsetting with OOB values

The change is identical in tibble 3.0.0 and we avoid a tibble warning. See https://cran.r-project.org/web/packages/tibble/vignettes/invariants.html for details (example is `df[10, ]`).

* Update min version requirement for tibble

* Modify creation of 0-row tibble

And roll back the requirement for the latest version of tibble.

* Use `if (interactive()) {` around examples that write files

This solution looks pretty nice in the internal help docs and in the pkgdown site (i.e., doesn't add cleanup lines in-between the example calls). Plus it shouldn't be flagged by a CRAN reviewer.

* Update NEWS.md