FIX: Fix LaTeX rendering of summary rows
Add 'boxh' formal argument to create_summary_rows().
'boxh' in create_summary_rows() allows the creation of 'default_vars' in a manner consistent with the method in summary_row_tags().
'default_vars' is used in the selection for 'summary_df'.

Wrapped 'visible_vars' with a call to unique(). Line 218 could add "rownames" to 'visible_vars' even if it already exists in 'default_vars' (from line 184). This keeps the table body from becoming misaligned.