Creat wrap environment functions
Instead of trying to add \begin{table} and \end{table} within various functions, I moved this step to the export.R as_latex function to create a wrapping environment that is either \begingroup-\endgroup for longtables or \begin{table}-\end{table} for floating tables. This will allow floating tables to be able to use fontsize changes and any other future changes that are meant to be added below the enclosing environment.