Add latex.tbl.pos option
Adds a latex.tbl.pos option to tab_options. The value here will be applied to latex output when a floating table is used to specify the floating position. It checks whether we are in a quarto environment to avoid messing up quarto output since floating position should be specified in tbl-pos chunk option in quarto docs.