Implement Shiny input.
I have a bunch of console.log()s in here right now as I debug. If you run Shiny.bindAll(); in Chrome devtools, this works, but not before then so I'm missing a step somewhere.

I'm not actually certain that the unsubscribe() method works, and I'm not yet certain how to test that.