Set default font size to NA so tables inherit document font
In LaTeX and HTML output formats, gt tables previously always set the
font size (see #1594). If the tables were being inserted into a
document (e.g., in a Quarto document), they did not inherit the font
size of the surrounding document, requiring the user to manually set the
size option for each table.

Instead, the default table.font.size is NA. Leaving it NA causes the
HTML and LaTeX outputs to emit no font size change, so the table is the
same size as the surrounding text in the document (or the browser's
default body font size). If the user manually sets the table.font.size,
it will be respected in both formats.

Fixes #1852.